move back to reg test suite
